WIP - Andorid launcher support. - Updates cargo.toml - Adds src/lib.rc and moves contents from src/main.rc to it. - Updates src/main.rc to call src/lib.rc
