+++ src-tauri/Cargo.toml
@@ -10,6 +10,18 @@ rust-version = "1.60"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+# # The `_lib` suffix may seem redundant but it is necessary
+# # to make the lib name unique and wouldn't conflict with the bin name.
+# # This seems to be only an issue on Windows, see https://github.com/rust-lang/cargo/issues/8519
+name = "bym_refitted_launcher_lib"
+path = "src/lib.rs"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[[bin]]
+name = "bym_refitted_launcher"
+path = "src/main.rs"
+
 [build-dependencies]
 tauri-build = { version = "2.0.4", features = [] }
 
@@ -17,13 +29,12 @@ tauri-build = { version = "2.0.4", features = [] }
 serde_json = "^1.0.135"
 serde = { version = "^1.0.217", features = ["derive"] }
 reqwest = { version = "^0.12.12", features = ["json"] }
-tauri = { version = "2.2.0", features = [
-    "devtools",
-] }
+tauri = { version = "2.2.0", features = ["devtools"] }
 tokio = { version = "^1.43", features = ["io-util", "fs"] }
 tauri-webview2 = "0.1.2"
 tauri-plugin-http = "2.2.0"
 tauri-plugin-process = "2.2.0"
+tauri-plugin-opener = "2"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
